remove items from contexts
