+++ src/extension.ts
@@ -1,20 +1,17 @@
 
 import * as vscode from "vscode";
-import * as models from "./models"
-import * as cmd from "./commands"
-import * as storage from './storage'
+import * as cmd from "./commands";
+import { ContextsProvider } from "./provider";
+import * as storage from './storage';
 
 
 export async function activate(context: vscode.ExtensionContext) {
-  const provider = new models.ContextsProvider(await storage.GetContexts());
+  const provider = new ContextsProvider(await storage.GetContexts());
   const treeView = vscode.window.createTreeView("ccs.contextExplorer", {
     treeDataProvider: provider,
     showCollapseAll: true,
   });
   const commands = cmd.Commands(provider, treeView)
-  provider.onDidChangeTreeData(() => {
-    storage.SaveContexts(provider.contexts())
-  });
   context.subscriptions.push(
     vscode.commands.registerCommand("ccs.createNewContext", commands.CreateNewContext),
     vscode.commands.registerCommand("ccs.renameContext", commands.RenameContext),
@@ -26,7 +23,18 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand("ccs.removeItemFromContext", commands.RemoveItemFromContext),
     vscode.commands.registerCommand("ccs.openItemInEditor", commands.OpenItemInEditor),
     vscode.commands.registerCommand("ccs.searchContext", commands.SearchContext),
-    vscode.commands.registerCommand("ccs.refresh", () => provider.refresh())
+    vscode.commands.registerCommand("ccs.refresh", () => provider.refresh()),
+    provider.onDidChangeTreeData(() => storage.SaveContexts(provider.contexts())),
+    vscode.workspace.onWillDeleteFiles((event) => {
+      event.waitUntil(
+        (async () => {
+          for (const f of event.files) {
+            await provider.removeResourceFromContexts(f);
+          }
+          provider.refresh();
+        })()
+      );
+    })
   );
 }
